pyproject.toml
[tool.poetry]
name = "newsletter-wizard"
version = "0.1.0"
description = ""
authors = ["Paul <paul.chaumeilv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
openai = "^1.58.1"
markdown = "^3.7"
requests = "^2.32.3"
schedule = "^1.2.2"
alembic = "^1.15.2"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
